upgrade RNG to mt19937_64 and seed types to uint64_t
- EAUtilRandom: std::mt19937 -> std::mt19937_64, uint32_t -> uint64_t in seed field, setSeed and getSeed
- EAConfiguration: uint32_t seed -> uint64_t
- EvolutionaryAlgorithm: uint32_t seed, baseSeed -> uint64_t
- EAStatistics: std::vector<long> seeds, long currentSeed -> uint64_t
- ea_main: long seed -> uint64_t, Experiment::maxruns int -> uint64_t

+++ base/EvolutionaryAlgorithm.h
@@ -33,11 +33,11 @@ namespace ea {
         /**
          * current seed for the RNG
          */
-        uint32_t seed;
+        uint64_t seed;
         /**
          * base seed for each batch of runs
          */
-        uint32_t baseSeed;
+        uint64_t baseSeed;
         /**
          * statistics of the EA
          */
